`tools/importer-rest-api-specs`: moving the Resource Identification logic into the Terraform package
