+++ tools/importer-rest-api-specs/pipeline/task_parse_data.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/discovery"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/parser"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/resources"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/terraform/resources"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/transformer"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/models"
 	"github.com/hashicorp/pandora/tools/sdk/config/definitions"
